feat: add realtime dropsonde plots
